requirements.txt
numpy
streamlit==1.9.0
tensorflow-cpu==2.9.0
opencv-python-headless==4.5.5.64
streamlit-webrtc==0.37.0
protobuf~=3.19.0
